# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Task Manager API",
      "version": "1.0.0",
      "description": "API for managing tasks"
    },
    "paths": {
      "/tasks": {
        "get": {
          "summary": "Get all tasks",
          "parameters": [
            {
              "name": "status",
              "in": "query",
              "description": "Filter tasks by status",
              "required": false,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "List of tasks",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": { "$ref": "#/components/schemas/Task" }
                  }
                }
              }
            }
          }
        },
        "post": {
          "summary": "Create a new task",
          "requestBody": {
            "description": "Task object to create",
            "required": true,
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          },
          "responses": {
            "201": { "description": "Task created successfully" }
          }
        }
      },
      "/tasks/{id}": {
        "get": {
          "summary": "Get a single task by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Task details",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/Task" }
                }
              }
            }
          }
        },
        "put": {
          "summary": "Update a task",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": { "type": "integer" }
            }
          ],
          "requestBody": {
            "description": "Task object to update",
            "required": true,
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Task" } }
            }
          },
          "responses": {
            "200": { "description": "Task updated successfully" }
          }
        },
        "delete": {
          "summary": "Delete a task",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": { "type": "integer" }
            }
          ],
          "responses": {
            "200": { "description": "Task deleted successfully" }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Task": {
          "type": "object",
          "properties": {
            "id": { "type": "integer" },
            "title": { "type": "string" },
            "description": { "type": "string" },
            "status": { "type": "string" },
            "due_date": { "type": "string", "format": "date" }
          }
        }
      }
    }
  }
  